Moves ChromeFlameGraphView into the new subfolder, to keep related classes together
