+++ vcperf.vcxproj
@@ -184,7 +184,7 @@
     <ClCompile Include="src\Analyzers\ExecutionHierarchy.cpp" />
     <ClCompile Include="src\Analyzers\ExpensiveTemplateInstantiationCache.cpp" />
     <ClCompile Include="src\Views\BuildExplorerView.cpp" />
-    <ClCompile Include="src\Views\ChromeFlameGraphView.cpp" />
+    <ClCompile Include="src\Views\ChromeFlameGraph\ChromeFlameGraphView.cpp" />
     <ClCompile Include="src\Views\ChromeFlameGraph\PackedProcessThreadRemapping.cpp" />
     <ClCompile Include="src\Views\FilesView.cpp" />
     <ClCompile Include="src\Views\FunctionsView.cpp" />
@@ -200,7 +200,7 @@
     <ClInclude Include="src\Analyzers\ExpensiveTemplateInstantiationCache.h" />
     <ClInclude Include="src\VcperfBuildInsights.h" />
     <ClInclude Include="src\Views\BuildExplorerView.h" />
-    <ClInclude Include="src\Views\ChromeFlameGraphView.h" />
+    <ClInclude Include="src\Views\ChromeFlameGraph\ChromeFlameGraphView.h" />
     <ClInclude Include="src\Views\ChromeFlameGraph\PackedProcessThreadRemapping.h" />
     <ClInclude Include="src\Views\FilesView.h" />
     <ClInclude Include="src\Views\FunctionsView.h" />
