+++ vcperf.vcxproj.filters
@@ -63,11 +63,11 @@
     <ClCompile Include="src\Analyzers\ExecutionHierarchy.cpp">
       <Filter>Source Files\Analyzers</Filter>
     </ClCompile>
-    <ClCompile Include="src\Views\ChromeFlameGraphView.cpp">
-      <Filter>Source Files\Views</Filter>
-    </ClCompile>
     <ClCompile Include="src\Views\ChromeFlameGraph\PackedProcessThreadRemapping.cpp">
       <Filter>Header Files\Views\ChromeFlameGraph</Filter>
+    </ClCompile>
+    <ClCompile Include="src\Views\ChromeFlameGraph\ChromeFlameGraphView.cpp">
+      <Filter>Source Files\Views\ChromeFlameGraph</Filter>
     </ClCompile>
   </ItemGroup>
   <ItemGroup>
@@ -110,11 +110,11 @@
     <ClInclude Include="src\Analyzers\ExecutionHierarchy.h">
       <Filter>Header Files\Analyzers</Filter>
     </ClInclude>
-    <ClInclude Include="src\Views\ChromeFlameGraphView.h">
-      <Filter>Header Files\Views</Filter>
-    </ClInclude>
     <ClInclude Include="src\Views\ChromeFlameGraph\PackedProcessThreadRemapping.h">
       <Filter>Source Files\Views\ChromeFlameGraph</Filter>
+    </ClInclude>
+    <ClInclude Include="src\Views\ChromeFlameGraph\ChromeFlameGraphView.h">
+      <Filter>Header Files\Views\ChromeFlameGraph</Filter>
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
